fix(types,core,chat,regression): fact args the kernel reads back as they were declared
The types corpus's P0 sweep baselined fact arguments whose Go values
contradict their Mangle Decl -- each a rule that could never fire, with
nothing to say so. Fixed at the sites outside lane A's packages, and
their baseline entries removed:

- edit_failed / delete_blocked reasons and campaign_intent_capture's
  autonomy level are /name; they were quoted strings. types.Atom (which
  had no caller) builds them. TestFileActionReasons_* reads them back
  through a real kernel.
- continuation_step is /number; it was passed float64.
- simulated_effect's argument list was rendered with %v ("[a b c]",
  boundaries lost); it is JSON now (effectArgsJSON).

Same class, found on the way:
- chat parseBool compared execution_result's Success against "true",
  but a /true reads back as "/true", so every delegated execution the
  chat summarised read as a failure. It uses types.ExtractBool.
- focus_resolution's confidence is an int64 percent; the chat read it as
  float64 (always 0) and multiplied by 100. It uses types.ArgFloat64.

Also wired: types.ClampLines -- `nerd regression run` prints a failed
task's output under the table, line-clamped head and tail, instead of
only "expected exit 0, got 1" -- and KernelTx.RetractPredicateSet, which
HydrateSessionContext now uses for the three predicates it replaces
whole. Removed: ExtractTime / ExtractDuration; the kernel reads TimeType
and DurationType back as strings neither parsed, and nothing asserts
either type.

Tests fail without the change except the hydration one, which pins the
predicate set's membership (a refactor).

+++ cmd/nerd/chat/model_update.go
@@ -1153,7 +1153,7 @@ func (m *Model) updateContinuationFacts() {
 	_ = m.kernel.Retract("continuation_step")
 	_ = m.kernel.Assert(core.Fact{
 		Predicate: "continuation_step",
-		Args:      []any{float64(m.continuationStep), float64(m.continuationTotal)},
+		Args:      []any{int64(m.continuationStep), int64(m.continuationTotal)},
 	})
 	// No step ceiling is asserted. Until 2026-09-18 max_continuation_steps(10)
 	// was written here and the continuation ended on it, which is a count as a
